requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu 
rouge_score 
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
torch
notebook
boto3
mypy-boto3-s3
python-box==6.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
evaluate
accelerate>=0.26.0
mlflow>=2.8.0
dagshub
pytest>=7.0.0
pytest-cov>=4.0.0